Refactor api structure
