+++ config/routes.rb
@@ -21,7 +21,9 @@ Rails.application.routes.draw do
   namespace :api, defaults: { format: :json } do
     namespace :v1 do
       namespace :admin do
-        resources :playlist_activities, only: [:update]
+        resources :students, only: [:update] do
+          patch '/playlist_activities/:id', controller: 'students/playlist_activities', action: 'update'
+        end
       end
     end
   end
